chore: rename and a bunch of stuff trying to get it to run its tests
