+++ app/initializers/ember-data-sails.js
@@ -1,4 +1,1 @@
-import EmberDataSailsInitializer from "@volldigital/ember-data-sails/initializers/ember-data-sails";
-
-export default EmberDataSailsInitializer;
-export var initialize = EmberDataSailsInitializer.initialize;
+export { default } from "voll-ember-data-sails/initializers/ember-data-sails";
