+++ app/serializers/sails.js
@@ -1,3 +1,1 @@
-import SailsSerializer from "@volldigital/ember-data-sails/serializers/sails";
-
-export default SailsSerializer;
+export { default } from "voll-ember-data-sails/serializers/sails";
