+++ config/environment.js
@@ -1,11 +1,12 @@
-'use strict';
+"use strict";
 
 module.exports = function (environment, appConfig) {
-  let level = environment === 'production' ? 'error' : 'warn';
+  let level = environment === "production" ? "error" : "warn";
   if (appConfig.SAILS_LOG_LEVEL) {
     level = appConfig.SAILS_LOG_LEVEL;
   }
   return {
-    LOG_LEVEL: level
+    modulePrefix: "voll-ember-data-sails",
+    LOG_LEVEL: level,
   };
 };
