+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@volldigital/ember-data-sails",
+  "name": "voll-ember-data-sails",
   "version": "1.0.0",
   "keywords": [
     "ember-addon"
@@ -27,15 +27,17 @@
     "test:all": "ember try:each"
   },
   "dependencies": {
-    "ember-cli-babel": "^7.26.11"
+    "ember-auto-import": "^2.6.3",
+    "ember-cli-babel": "^7.26.11",
+    "lodash": "^4.17.21"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.22.6",
     "@ember/optional-features": "^1.3.0",
+    "@ember/test-helpers": "^2.9.4",
     "@glimmer/component": "^1.0.0",
     "@glimmer/tracking": "^1.0.0",
     "broccoli-asset-rev": "^3.0.0",
-    "ember-auto-import": "^2.6.3",
     "ember-cli": "~4.0.1",
     "ember-cli-dependency-checker": "^3.3.2",
     "ember-cli-htmlbars": "^6.2.0",
@@ -49,14 +51,14 @@
     "ember-resolver": "^8.1.0",
     "ember-source": "~4.0.1",
     "ember-source-channel-url": "^3.0.0",
+    "ember-template-lint": "^5.11.0",
     "ember-try": "^2.0.0",
     "eslint": "^8.44.0",
     "eslint-plugin-ember": "^11.9.0",
     "eslint-plugin-node": "^11.1.0",
-    "ember-template-lint": "^5.11.0",
     "loader.js": "^4.7.0",
-    "lodash": "^4.17.21",
-    "qunit-dom": "^2.0.0"
+    "qunit-dom": "^2.0.0",
+    "webpack": "^5.88.1"
   },
   "engines": {
     "node": "6.* || 8.* || >= 10.*"
