+++ tests/dummy/app/adapters/application.js
@@ -1,4 +1,4 @@
-import SailsSocketAdapter from "@volldigital/ember-data-sails/adapters/sails-socket";
+import SailsSocketAdapter from "voll-ember-data-sails/adapters/sails-socket";
 
 export default class SailsApplicationAdapter extends SailsSocketAdapter {
   defaultSerializer = "-rest";
