+++ tests/dummy/app/services/sails-socket.js
@@ -1,3 +1,3 @@
-import _SailsSocketService from "@volldigital/ember-data-sails/services/sails-socket";
+import _SailsSocketService from "voll-ember-data-sails/services/sails-socket";
 
 export default class SailsSocketService extends _SailsSocketService {}
