+++ tests/helpers/adapter.js
@@ -1,6 +1,6 @@
 import { Promise as EmberPromise } from "rsvp";
-import SailsSocketAdapter from "@volldigital/ember-data-sails/adapters/sails-socket";
-import SailsRESTAdapter from "@volldigital/ember-data-sails/adapters/sails-rest";
+import SailsSocketAdapter from "voll-ember-data-sails/adapters/sails-socket";
+import SailsRESTAdapter from "voll-ember-data-sails/adapters/sails-rest";
 
 export function adapterCall(adapter, returnPromise, method) {
   const args = [].slice.call(arguments, 2);
